Delete works for modules.
Also fixed a bug in days error response. When getting a day by ID we checked to see if it exists. If our query returned None. We would pass the None to the error message rather than the bad ID.

Also need to think about handling file deletions as a result of classroom, unit, module, and day deletions.